+++ src/useRouteQueryParams/index.ts
@@ -1,3 +1,2 @@
-export * from './useRouteQueryParam'
 export * from './useRouteQueryParams'
 export * from './formats'
